huggingface-space/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
marker-pdf==0.2.17
torch==2.1.0
transformers==4.36.2
Pillow==10.1.0
